Marking nodejs errors as test failures
While the concept of a "failure" in jUnit is usually referring to the failure of test assertions,
for whatever reason Jenkins does not mark a build as "failed" if various tests error out.

Using the "failure" mark will break the builds as expected, even if it doesn't strictly follow
the convention set out by jUnit.